# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Water-id Translator Documentation",
        "description": "L5 Swagger OpenApi description",
        "contact": {
            "email": "admin@water-id.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "2.0.0"
    },
    "servers": [
        {
            "url": "http://translator.local/api/v1",
            "description": "Water-id Translator API Server"
        }
    ],
    "paths": {
...
        "/phrases": {
            "get": {
                "tags": [
                    "Phrases"
                ],
                "summary": "Get list of phrases",
                "description": "Returns list of phrases",
                "operationId": "getPhrasesList",
                "parameters": [
                    {
                        "name": "page",
                        "in": "query",
                        "description": "Page number",
                        "required": false,
                        "schema": {
                            "type": "integer"
                        },
                        "example": 1
                    },
                    {
                        "name": "perPage",
                        "in": "query",
                        "description": "Number of item per page",
                        "schema": {
                            "type": "integer"
                        },
                        "example": 20
                    },
                    {
                        "name": "sortBy",
                        "in": "query",
                        "description": "Column name to sort by",
                        "schema": {
                            "type": "string"
                        },
                        "example": "id"
                    },
                    {
                        "name": "sortDesc",
                        "in": "query",
                        "description": "Column name to sort descending",
                        "required": true,
                        "schema": {
                            "type": "boolean"
                        },
                        "example": true
                    },
                    {
                        "name": "q",
                        "in": "query",
                        "description": "Search Query",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "phrase",
                        "in": "query",
                        "description": "Phrase",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "category",
                        "in": "query",
                        "description": "Category name",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "phraseCategoryId",
                        "in": "query",
                        "description": "Phrase Category Id",
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "phraseCategory",
                        "in": "query",
                        "description": "Phrase category name",
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/PhraseResource"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ApiUnAuthException"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Invalid scope(s) provided.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ApiAccessDeniedException"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "passport": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Phrases"
                ],
                "summary": "Store new phrase",
                "description": "Returns phrase data",
                "operationId": "storePhrase",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/StorePhraseRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/PhraseOut"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ApiRequestException"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ApiUnAuthException"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Invalid scope(s) provided.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ApiAccessDeniedException"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Unprocessable Entity",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ApiUnprocessableEntityException"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "passport": [
                            "*"
                        ]
                    }
                ]
            }
        },
        "/phrases/{id}": {
            "get": {
                "tags": [
                    "Phrases"
                ],
                "summary": "Get phrase information",
                "description": "Returns phrase data",
                "operationId": "getPhraseById",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Phrase id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/PhraseOut"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ApiRequestException"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ApiUnAuthException"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Invalid scope(s) provided.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ApiAccessDeniedException"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Resource Not Found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ApiNotFoundException"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "passport": [
                            "*"
                        ]
                    }
                ]
            },
            "put": {
                "tags": [
                    "Phrases"
                ],
                "summary": "Update existing phrase",
                "description": "Returns updated phrase data",
                "operationId": "updatePhrase",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Phrase id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/UpdatePhraseRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "202": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/PhraseOut"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ApiRequestException"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ApiUnAuthException"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Invalid scope(s) provided.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ApiAccessDeniedException"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Resource Not Found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ApiNotFoundException"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "passport": [
                            "*"
                        ]
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Phrases"
                ],
                "summary": "Delete existing phrase",
                "description": "Deletes a record and returns no content",
                "operationId": "deletePhrase",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Phrase id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ApiUnAuthException"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Invalid scope(s) provided.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ApiAccessDeniedException"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Resource Not Found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ApiNotFoundException"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "passport": [
                            "*"
                        ]
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
            "ApiAccessDeniedException": {
                "properties": {
                    "message": {
                        "description": "The err message",
                        "type": "string",
                        "example": "AccessDenied"
                    }
                },
                "type": "object"
            },
            "ApiNotFoundException": {
                "properties": {
                    "message": {
                        "description": "The err message",
                        "type": "string",
                        "example": "Not Found"
                    }
                },
                "type": "object"
            },
...
            "ApiRequestException": {
                "properties": {
                    "message": {
                        "description": "The err message",
                        "type": "string",
                        "example": "Bad Request"
                    }
                },
                "type": "object"
            },
...
            "ApiUnAuthException": {
                "properties": {
                    "message": {
                        "description": "The err message",
                        "type": "string",
                        "example": "Unauthenticated"
                    }
                },
                "type": "object"
            },
            "ApiUnprocessableEntityException": {
                "properties": {
                    "message": {
                        "description": "The err message",
                        "type": "string",
                        "example": "Unprocessable Entity"
                    }
                },
                "type": "object"
            },
...
            "StorePhraseRequest": {
                "title": "Store Phrase request",
                "description": "Store Phrase request body data",
                "required": [
                    "phrase"
                ],
                "properties": {
                    "base_id": {
                        "title": "Base ID",
                        "description": "Base ID",
                        "type": "integer",
                        "format": "int64",
                        "example": 1
                    },
                    "phrase": {
                        "title": "Phrase",
                        "description": "Phrase",
                        "type": "string",
                        "example": "A new phrase"
                    },
                    "category_name": {
                        "title": "Category Name",
                        "description": "Category Name",
                        "type": "string",
                        "example": "A category name"
                    },
                    "phrase_category_id": {
                        "title": "Phrase Category ID",
                        "description": "Phrase Category ID",
                        "type": "integer",
                        "format": "int64",
                        "example": 1
                    }
                },
                "type": "object"
            },
...
            "PhraseResource": {
                "title": "PhraseResource",
                "description": "Phrase resource",
                "properties": {
                    "data": {
                        "title": "Data",
                        "description": "Data wrapper",
                        "type": "array",
                        "items": {
                            "ref": null
                        }
                    }
                },
                "type": "object",
                "xml": {
                    "name": "PhraseResource"
                }
            },
...
            "PhraseOut": {
                "title": "PhraseOut",
                "description": "PhraseOut model",
                "properties": {
                    "id": {
                        "title": "ID",
                        "description": "ID",
                        "type": "integer",
                        "format": "int64",
                        "example": 1
                    },
                    "base_id": {
                        "title": "Base ID",
                        "description": "Base ID",
                        "type": "integer",
                        "format": "int64",
                        "example": 1
                    },
                    "phrase": {
                        "title": "Phrase",
                        "description": "Phrase",
                        "type": "string",
                        "example": "A new phrase"
                    },
                    "category_name": {
                        "title": "Category Name",
                        "description": "Category Name",
                        "type": "string",
                        "example": "A category name"
                    },
                    "phrase_category_id": {
                        "title": "Phrase Category ID",
                        "description": "Phrase Category ID",
                        "type": "integer",
                        "format": "int64",
                        "example": 1
                    },
                    "translations_count": {
                        "title": "Translations Count",
                        "description": "Translations Count",
                        "type": "integer",
                        "format": "int64",
                        "example": 1
                    },
                    "created_at": {
                        "title": "Created at",
                        "description": "Created at",
                        "type": "string",
                        "format": "datetime",
                        "example": "2020-01-27 17:50:45"
                    },
                    "updated_at": {
                        "title": "Updated at",
                        "description": "Updated at",
                        "type": "string",
                        "format": "datetime",
                        "example": "2020-01-27 17:50:45"
                    },
                    "deleted_at": {
                        "title": "Deleted at",
                        "description": "Deleted at",
                        "type": "string",
                        "format": "datetime",
                        "example": "2020-01-27 17:50:45"
                    },
                    "phraseCategory": {
                        "$ref": "#/components/schemas/PhraseCategoryOut"
                    },
                    "category": {
                        "$ref": "#/components/schemas/CategoryOut"
                    }
                },
                "type": "object",
                "xml": {
                    "name": "PhraseOut"
                }
            },
...
        },
        "securitySchemes": {
            "passport": {
                "type": "apiKey",
                "description": "Enter \"Bearer + [space] + [token]\" in the text input below(Value).",
                "example": "Bearer 1qaz2wsx3edc",
                "in": "header",
                "name": "Authorization"
            }
        }
    },
...
}